% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{analyze}
\alias{analyze}
\title{Analyze
Perform the type checks for a given file}
\usage{
analyze(filepath)
}
\arguments{
\item{filepath}{Path to the analyzed file}
}
\value{
Boolean whether the check passes
}
\description{
Analyze
Perform the type checks for a given file
}
